+++ ChatApp/Model/Message.swift
@@ -10,5 +10,5 @@ import Foundation
 struct Message {
     let userID: Int
     let text: String
-    let date: Date
+    let date: String
 }
